+++ src/main/java/com/gotocompany/depot/maxcompute/converter/mapper/ProtoPrimitiveDataTypeMapperFactory.java
@@ -1,8 +1,15 @@
-package com.gotocompany.depot.maxcompute.converter.strategy;
+package com.gotocompany.depot.maxcompute.converter.mapper;
 
 import com.gotocompany.depot.config.MaxComputeSinkConfig;
+import com.gotocompany.depot.maxcompute.converter.mapper.casted.DoubleToDecimalDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.casted.FloatToDecimalDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.casted.IntegerToBigintDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.noncasted.DoubleDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.noncasted.FloatDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.noncasted.IntegerDataTypeMapper;
+import com.gotocompany.depot.maxcompute.converter.mapper.noncasted.NonNumericDataTypeMapper;
 
-public class ProtoPrimitiveDataTypeMappingStrategyFactory {
+public class ProtoPrimitiveDataTypeMapperFactory {
 
     public static ProtoPrimitiveDataTypeMapper createPrimitiveProtobufMappingStrategy(MaxComputeSinkConfig maxComputeSinkConfig) {
         ProtoPrimitiveDataTypeMapper baseProtoPrimitiveDataTypeMapper = new NonNumericDataTypeMapper();
