+++ src/main/java/com/gotocompany/depot/maxcompute/converter/mapper/casted/DoubleToDecimalDataTypeMapper.java
@@ -1,4 +1,4 @@
-package com.gotocompany.depot.maxcompute.converter.strategy;
+package com.gotocompany.depot.maxcompute.converter.mapper.casted;
 
 import com.aliyun.odps.type.TypeInfo;
 import com.aliyun.odps.type.TypeInfoFactory;
@@ -6,6 +6,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.Descriptors;
 import com.gotocompany.depot.config.MaxComputeSinkConfig;
 import com.gotocompany.depot.exception.InvalidMessageException;
+import com.gotocompany.depot.maxcompute.converter.mapper.ProtoPrimitiveDataTypeMapper;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
@@ -13,40 +14,36 @@ import java.math.RoundingMode;
 import java.util.Map;
 import java.util.function.Function;
 
-import static com.google.protobuf.Descriptors.FieldDescriptor.Type.FLOAT;
+public class DoubleToDecimalDataTypeMapper implements ProtoPrimitiveDataTypeMapper {
 
-public class FloatToDecimalDataTypeMapper implements ProtoPrimitiveDataTypeMapper {
-
-    private final int scale;
     private final int precision;
+    private final int scale;
     private final RoundingMode roundingMode;
 
-    public FloatToDecimalDataTypeMapper(MaxComputeSinkConfig maxComputeSinkConfig) {
-        this.scale = maxComputeSinkConfig.getProtoFloatToDecimalScale();
-        this.precision = maxComputeSinkConfig.getProtoFloatToDecimalPrecision();
+    public DoubleToDecimalDataTypeMapper(MaxComputeSinkConfig maxComputeSinkConfig) {
+        this.precision = maxComputeSinkConfig.getProtoDoubleToDecimalPrecision();
+        this.scale = maxComputeSinkConfig.getProtoDoubleToDecimalScale();
         this.roundingMode = maxComputeSinkConfig.getDecimalRoundingMode();
     }
 
     @Override
     public Map<Descriptors.FieldDescriptor.Type, TypeInfo> getProtoTypeMap() {
         return ImmutableMap.<Descriptors.FieldDescriptor.Type, TypeInfo>builder()
-                .put(FLOAT, TypeInfoFactory.getDecimalTypeInfo(precision, scale))
+                .put(Descriptors.FieldDescriptor.Type.DOUBLE, TypeInfoFactory.getDecimalTypeInfo(precision, scale))
                 .build();
     }
 
     @Override
     public Map<Descriptors.FieldDescriptor.Type, Function<Object, Object>> getProtoPayloadMapperMap() {
         return ImmutableMap.<Descriptors.FieldDescriptor.Type, Function<Object, Object>>builder()
-                .put(FLOAT, object -> isValid((float) object))
+                .put(Descriptors.FieldDescriptor.Type.DOUBLE, value -> isValid((double) value))
                 .build();
     }
 
-    private BigDecimal isValid(float value) {
-        if (!Float.isFinite(value)) {
-            throw new InvalidMessageException("Invalid float value: " + value);
+    private BigDecimal isValid(double value) {
+        if (!Double.isFinite(value)) {
+            throw new InvalidMessageException("Invalid double value: " + value);
         }
-        return new BigDecimal(Float.toString(value), new MathContext(precision))
-                .setScale(scale, roundingMode);
+        return new BigDecimal(String.valueOf(value), new MathContext(precision)).setScale(scale, roundingMode);
     }
-
 }
