+++ src/main/java/com/gotocompany/depot/maxcompute/converter/mapper/noncasted/DoubleDataTypeMapper.java
@@ -1,36 +1,36 @@
-package com.gotocompany.depot.maxcompute.converter.strategy;
+package com.gotocompany.depot.maxcompute.converter.mapper.noncasted;
 
 import com.aliyun.odps.type.TypeInfo;
 import com.aliyun.odps.type.TypeInfoFactory;
 import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.Descriptors;
 import com.gotocompany.depot.exception.InvalidMessageException;
+import com.gotocompany.depot.maxcompute.converter.mapper.ProtoPrimitiveDataTypeMapper;
 
 import java.util.Map;
 import java.util.function.Function;
 
-import static com.google.protobuf.Descriptors.FieldDescriptor.Type.FLOAT;
+import static com.google.protobuf.Descriptors.FieldDescriptor.Type.DOUBLE;
 
-
-public class FloatDataTypeMapper implements ProtoPrimitiveDataTypeMapper {
+public class DoubleDataTypeMapper implements ProtoPrimitiveDataTypeMapper {
 
     @Override
     public Map<Descriptors.FieldDescriptor.Type, TypeInfo> getProtoTypeMap() {
         return ImmutableMap.<Descriptors.FieldDescriptor.Type, TypeInfo>builder()
-                .put(FLOAT, TypeInfoFactory.FLOAT)
+                .put(Descriptors.FieldDescriptor.Type.DOUBLE, TypeInfoFactory.DOUBLE)
                 .build();
     }
 
     @Override
     public Map<Descriptors.FieldDescriptor.Type, Function<Object, Object>> getProtoPayloadMapperMap() {
         return ImmutableMap.<Descriptors.FieldDescriptor.Type, Function<Object, Object>>builder()
-                .put(FLOAT, object -> isValid((float) object))
+                .put(DOUBLE, object -> isValid((double) object))
                 .build();
     }
 
-    private static float isValid(float value) {
-        if (!Float.isFinite(value)) {
-            throw new InvalidMessageException("Invalid float value: " + value);
+    private static double isValid(double value) {
+        if (!Double.isFinite(value)) {
+            throw new InvalidMessageException("Invalid double value: " + value);
         }
         return value;
     }
