+++ src/test/java/com/gotocompany/depot/maxcompute/converter/mapper/noncasted/FloatDataTypeMapperTest.java
@@ -1,4 +1,4 @@
-package com.gotocompany.depot.maxcompute.converter.strategy;
+package com.gotocompany.depot.maxcompute.converter.mapper.noncasted;
 
 import com.aliyun.odps.OdpsType;
 import com.aliyun.odps.type.TypeInfo;
@@ -8,7 +8,7 @@ import org.junit.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class FloatDataTypeMappingStrategyTest {
+public class FloatDataTypeMapperTest {
 
     private final FloatDataTypeMapper floatPrimitiveProtobufMappingStrategy = new FloatDataTypeMapper();
